pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "torch-tensorcore-speedkit"
version = "0.1.0"
description = "Auditable PyTorch Tensor Core optimization via auto-detection"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "torch-tensorcore-speedkit contributors"}]
keywords = ["pytorch", "tensor-cores", "performance", "tf32", "bf16", "optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "PyYAML>=6.0.1",
  "tqdm>=4.66",
  "numpy>=1.24",
]

[project.optional-dependencies]
vision = ["torchvision>=0.15"]
dev = ["pytest>=7.4", "ruff>=0.6.0", "mypy>=1.7"]

[project.urls]
Homepage = "https://github.com/yourusername/torch-tensorcore-speedkit"
Repository = "https://github.com/yourusername/torch-tensorcore-speedkit"
Issues = "https://github.com/yourusername/torch-tensorcore-speedkit/issues"
Changelog = "https://github.com/yourusername/torch-tensorcore-speedkit/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["torch_speedkit"]

